+++ assets/sidebar.js
@@ -2,17 +2,19 @@ const tsVscode = acquireVsCodeApi();
 
 var regexOptionG = document.getElementById('regex-option-g');
 var regexOptionI = document.getElementById('regex-option-i');
-var regexOptionM = document.getElementById('regex-option-m');
+var regexOptionM = document.getElementById('regex-option-m')
+var regexOptionU = document.getElementById('regex-option-u')
+var regexOptionS = document.getElementById('regex-option-s');
 var regexOptionY = document.getElementById('regex-option-y');
 var regexInput = document.getElementById('regex-input');
-
-var webviewError = document.getElementById('webview-error');
 var resultTree = document.getElementById('result-tree');
 
 regexInput.addEventListener('input', update);
 regexOptionG.addEventListener('change', update);
 regexOptionI.addEventListener('change', update);
 regexOptionM.addEventListener('change', update);
+regexOptionU.addEventListener('change', update);
+regexOptionS.addEventListener('change', update);
 regexOptionY.addEventListener('change', update);
 
 let updateTimeout = null;
@@ -23,15 +25,16 @@ function update() {
   var options = (regexOptionG.checked ? 'g' : '') +
                 (regexOptionI.checked ? 'i' : '') +
                 (regexOptionM.checked ? 'm' : '') +
+                (regexOptionU.checked ? 'u' : '') +
+                (regexOptionS.checked ? 's' : '') +
                 (regexOptionY.checked ? 'y' : '');
 
   updateTimeout = setTimeout(function() {
-    webviewError.innerHTML = '';
 
     if(options == '') {
       resultTree.innerHTML = '';
       tsVscode.postMessage({type: 'clearRegex'});
-      webviewError.innerHTML = 'Select at least on checkbox option';
+      tsVscode.postMessage({type: 'onError', message: 'Select at least on checkbox option'});
       return; 
     }
 
@@ -40,31 +43,9 @@ function update() {
       tsVscode.postMessage({type: 'clearRegex'});
       return; 
     }
-  
-    var err = checkRegExp(value, options);
-    if(err != null) {
-      webviewError.innerHTML = err;
-      return;
-    }
 
     tsVscode.postMessage({type: 'updateRegex', regex: value, options: options});
   }, 500);
-}
-
-function checkRegExp(regexString, options) {
-  try {
-    return new Function(`
-      "use strict";
-      try {
-          new RegExp('${regexString}', '${options}');
-          return null;
-      } catch (e) {
-          return e;
-      }
-    `)();
-  } catch(e) {
-      return e;
-  }
 }
 
 window.addEventListener('message', event => {