+++ src/regex-handler.ts
@@ -7,14 +7,20 @@ import { SidebarProvider } from './webview/sidebar';
  * @implements {vscode.WebviewViewProvider}
  */
 export class RegexHandler {
+  lock:boolean;
+  regexString:string;
   regex?:RegExp; 
   decorationType:vscode.TextEditorDecorationType; 
 
   /**
    * Constructor of the class.
    * @param {vscode.ExtensionContext}
+   * @param {SidebarProvider}
    */ 
    constructor(context:vscode.ExtensionContext, private sidebarProvider:SidebarProvider) {
+    this.lock = false;
+    this.regexString = '';
+
     const configMap = config.ConfigurationToMap();
     this.decorationType = vscode.window.createTextEditorDecorationType({
       backgroundColor: configMap.get('backgroundColor'),
@@ -69,42 +75,54 @@ export class RegexHandler {
    * @param {vscode.TextDocument | null}
    */ 
   public onDocumentChanged(document:vscode.TextDocument | null) {
-    if(this.sidebarProvider.webviewView == undefined || !this.sidebarProvider.webviewView.visible) {
-      vscode.window.visibleTextEditors.map( editor => {
-        editor.setDecorations(this.decorationType, []);
-      });
-      return;
-    }
+    if(this.lock) return;
+    this.lock = true;
     
-    var documentMatches: RegexMatch[] = []; 
+    try {
+      if(this.sidebarProvider.webviewView == undefined || !this.sidebarProvider.webviewView.visible) {
+        vscode.window.visibleTextEditors.map( editor => {
+          editor.setDecorations(this.decorationType, []);
+        });
 
-    if(this.regex != undefined) {
-      vscode.window.visibleTextEditors.map( editor => {
-        var decorationsArray: vscode.DecorationOptions[] = [];
-
-        if(document == null || document == editor.document) {
-          var documentMatch = new RegexMatch(editor.document.fileName);
-          const text = editor.document.getText();
-          var match:RegExpExecArray|null;
-
-          // I have checked for undefined before but the linter cannot perceive it...
-          while(this.regex != undefined && (match = this.regex.exec(text)) !== null && match.index !== undefined) {
-            var range = new vscode.Range(
-              editor.document.positionAt(match.index),
-              editor.document.positionAt(match.index + match[0].length)
-            );
-
-            documentMatch.matches.push(match);
-            decorationsArray.push({ range });
+        throw 'clean stop';
+      }
+      
+      var documentMatches: RegexMatch[] = []; 
+  
+      if(this.regex != undefined) {
+        vscode.window.visibleTextEditors.map( editor => {
+          var decorationsArray: vscode.DecorationOptions[] = [];
+  
+          if(document == null || document == editor.document) {
+            var documentMatch = new RegexMatch(editor.document.fileName);
+            const text = editor.document.getText();
+            var match:RegExpExecArray|null;
+  
+            // I have checked for undefined before but the linter cannot perceive it...
+            while(this.regex != undefined && (match = this.regex.exec(text)) !== null && match.index !== undefined) {
+              var range = new vscode.Range(
+                editor.document.positionAt(match.index),
+                editor.document.positionAt(match.index + match[0].length)
+              );
+  
+              documentMatch.matches.push(match);
+              decorationsArray.push({ range });
+            }
+  
+            documentMatches.push(documentMatch);
+            editor.setDecorations(this.decorationType, decorationsArray);
           }
-
-          documentMatches.push(documentMatch);
-          editor.setDecorations(this.decorationType, decorationsArray);
-        }
-      });
-
-      this.sidebarProvider.webviewView.webview.postMessage({type: 'documentMatches', data: documentMatches});
+        });
+  
+        this.sidebarProvider.webviewView.webview.postMessage({type: 'documentMatches', data: documentMatches});
+      }
+    } catch(err) {
+      if(err.message != 'clean stop') {
+        vscode.window.showErrorMessage(err.message);
+      }
     }
+
+    this.lock = false;
   }
 
   /**
@@ -124,10 +142,11 @@ export class RegexHandler {
   public Set(regexString:string, options:string) {
     try {
       this.regex = new RegExp(regexString, options);
+      this.regexString = regexString;
       
       this.onDocumentChanged(null);
-    } catch(e) {
-      vscode.window.showErrorMessage(`Regex "${regexString}" is invalid`);
+    } catch(err) {
+      vscode.window.showErrorMessage(`Regex "${regexString}" is invalid\n${err.message}`);
       this.regex = undefined;
     }
   }