+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
 ## [Unreleased]
-
-- Initial release
+- Add explain webview
+- Move error checking to extension and add lock (changes prevent crash bug)
+- Add issuer and feature request templates
+- Change input placeholder and update roadmap
+- Add flag u and s
+- Initial Commit